+++ server.js
@@ -13,8 +13,6 @@ var search = require('./models/search')
 
 var app = express();
 
-console.log(process.env.NODE_ENV)
-
 app.set('port', process.env.PORT || 3000);
 app.use(express.static(path.join(__dirname, 'public')));
 app.use(bodyParser.json());