+++ CHANGELOG.md
@@ -4,15 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.2] - 2017-02-08
 ### Added
 - Search form.
 - Search api.
 - Implement search function.
 - Env based db configuration.
+- Pass selected to search field.
 
 ### Changed
 - Align of time selector dropmenu.
+- Adjust query precision
 
 ### Fixed
 - The course list is disappeared.